+++ web_gpu_app.cpp
@@ -17,17 +17,6 @@
 
 namespace {
 
-void OnDeviceError(WGPUErrorType type, const char* message, void* userdata) {
-  std::cout << "Dawn error: " << message << std::endl;
-  exit(0);
-}
-
-void OnDeviceLost(WGPUDeviceLostReason reason, char const* message, void* userdata) {
-  std::cout << "Dawn device lost: " << message << std::endl;
-  std::cout << "Reason: " << reason << std::endl;
-  std::cout << "Device: " << userdata << std::endl;
-}
-
 #if defined(__EMSCRIPTEN__)
 void EmscriptenMainLoop(void* app) { reinterpret_cast<App*>(app)->Render(); }
 #endif
@@ -36,22 +25,29 @@ web_gpu_app::App* AppFromWindow(GLFWwindow* window) {
   return reinterpret_cast<web_gpu_app::App*>(glfwGetWindowUserPointer(window));
 }
 
-}  // OnGlfwResize
+void OnGlfwResize(GLFWwindow* window, int width, int height) {
+  AppFromWindow(window)->OnResize(width, height);
+}
+
+void OnGlfwSetCursorPos(GLFWwindow* window, double xpos, double ypos) {
+  AppFromWindow(window)->OnMouseMove(xpos, ypos);
+}
+
+void OnGlfwSetMouseButton(GLFWwindow* window, int button, int action, int mods) {
+  AppFromWindow(window)->OnMouseButton(button, action, mods);
+}
+
+void OnGlfwScroll(GLFWwindow* window, double x_offset, double y_offset) {
+  AppFromWindow(window)->OnScroll(x_offset, y_offset);
+}
+
+}  // namespace
 
 namespace web_gpu_app {
 
-const char simple_shader_code[] = R"(
-    @vertex fn vertexMain(@builtin(vertex_index) i : u32) ->
-      @builtin(position) vec4f {
-        const pos = array(vec2f(0, 1), vec2f(-1, -1), vec2f(1, -1));
-        return vec4f(pos[i], 0, 1);
-    }
-    @fragment fn fragmentMain() -> @location(0) vec4f {
-        return vec4f(0.1, 0.4, 0, 1);
-    }
-)";
-
 App::App() {
+  window_ = SetupGlfwWindow(GetAppName().c_str(), this);
+  renderer_ = std::make_unique<WebGpuRenderer>(window_);
   // Imgui
   SetupUi();
   MainLoop();
@@ -68,74 +64,9 @@ void App::MainLoop() {
 #else
   while (!glfwWindowShouldClose(window_)) {
     glfwPollEvents();
-    Render();
-    swap_chain_.Present();
+    renderer_->Render({});
   }
 #endif
-}
-
-void App::Render() {
-  wgpu::RenderPassColorAttachment attachment{.view = swap_chain_.GetCurrentTextureView(),
-                                             .loadOp = wgpu::LoadOp::Clear,
-                                             .storeOp = wgpu::StoreOp::Store};
-
-  wgpu::RenderPassDepthStencilAttachment depthStencilAttachment;
-  depthStencilAttachment.view = depth_texture_view_;
-  depthStencilAttachment.depthClearValue = 1.0f;
-  depthStencilAttachment.depthLoadOp = wgpu::LoadOp::Clear;
-  depthStencilAttachment.depthStoreOp = wgpu::StoreOp::Store;
-  depthStencilAttachment.depthReadOnly = false;
-  depthStencilAttachment.stencilClearValue = 0;
-#ifdef WEBGPU_BACKEND_WGPU
-  depthStencilAttachment.stencilLoadOp = wgpu::LoadOp::Clear;
-  depthStencilAttachment.stencilStoreOp = wgpu::StoreOp::Store;
-#else
-  depthStencilAttachment.stencilLoadOp = wgpu::LoadOp::Undefined;
-  depthStencilAttachment.stencilStoreOp = wgpu::StoreOp::Undefined;
-#endif
-  depthStencilAttachment.stencilReadOnly = true;
-
-  wgpu::RenderPassDescriptor renderpass{.colorAttachmentCount = 1,
-                                        .colorAttachments = &attachment,
-                                        .depthStencilAttachment = &depthStencilAttachment};
-
-  wgpu::CommandEncoder encoder = device_.CreateCommandEncoder();
-  wgpu::RenderPassEncoder pass = encoder.BeginRenderPass(&renderpass);
-  pass.SetPipeline(render_pipeline_);
-  pass.Draw(3);
-  RenderUi(pass);
-  pass.End();
-  wgpu::CommandBuffer commands = encoder.Finish();
-  device_.GetQueue().Submit(1, &commands);
-  device_.Tick();
-}
-
-void App::RenderUi(wgpu::RenderPassEncoder render_pass) {
-  ImGui_ImplWGPU_NewFrame();
-  ImGui_ImplGlfw_NewFrame();
-  ImGui::NewFrame();
-
-  ImGui::ShowDemoWindow();
-
-  ImGui::EndFrame();
-  ImGui::Render();
-  ImGui_ImplWGPU_RenderDrawData(ImGui::GetDrawData(), render_pass.Get());
-}
-
-void App::OnGlfwResize(GLFWwindow* window, int width, int height) {
-  AppFromWindow(window)->OnResize(width, height);
-}
-
-void App::OnGlfwSetCursorPos(GLFWwindow* window, double xpos, double ypos) {
-  AppFromWindow(window)->OnMouseMove(xpos, ypos);
-}
-
-void App::OnGlfwSetMouseButton(GLFWwindow* window, int button, int action, int mods) {
-  AppFromWindow(window)->OnMouseButton(button, action, mods);
-}
-
-void App::OnGlfwScroll(GLFWwindow* window, double x_offset, double y_offset) {
-  AppFromWindow(window)->OnScroll(x_offset, y_offset);
 }
 
 GLFWwindow* App::SetupGlfwWindow(const char* title, void* user_pointer) {
@@ -163,18 +94,12 @@ void App::SetupUi() {
   ImGui::CreateContext();
   ImGui::GetIO();
   ImGui_ImplGlfw_InitForOther(window_, true);
-  ImGui_ImplWGPU_Init(device_.Get(), 3, WGPUTextureFormat_BGRA8Unorm,
+  ImGui_ImplWGPU_Init(renderer_->GetDevice(), 3, WGPUTextureFormat_BGRA8Unorm,
                       WGPUTextureFormat_Depth24Plus);
   SetUiThemeDark();
 }
 
-void App::OnResize(int width, int height) {
-  width_ = width;
-  height_ = height;
-  swap_chain_ = SetupSwapChain(surface_, device_, width_, height_);
-  depth_texture_ = SetupDepthTexture(device_, depth_texture_format_, width_, height_);
-  depth_texture_view_ = SetupDepthTextureView(depth_texture_, depth_texture_format_);
-}
+void App::OnResize(int width, int height) { renderer_->OnResize(width, height); }
 
 void App::OnMouseMove(double xpos, double ypos) {}
 void App::OnMouseButton(int button, int action, int mods) {
